Remplacer usecases-ban.json par l'appel api + cache next de 24h
