+++ .env.default
@@ -30,6 +30,13 @@ NEXT_LAST_UPDATED_DATE_ADDOCK_URL='https://api-adresse.data.gouv.fr/info.txt'
 # Geocodeur Géoplateforme
 NEXT_PUBLIC_API_ADRESSE=https://data.geopf.fr/geocodage/
 
+# API Grist pour la page usages
+NEXT_PUBLIC_GRIST_API_URL=https://grist.numerique.gouv.fr/api
+NEXT_PUBLIC_GRIST_DOC_ID=
+NEXT_PUBLIC_GRIST_TABLE_ID=
+NEXT_PUBLIC_GRIST_BAN_ORGANIZATION_ID=
+NEXT_PUBLIC_GRIST_WORKSPACE_ID=
+NEXT_PUBLIC_GRIST_API_TOKEN=
 
 # -------------------------------------
 # --- Utilisation appel externe Bal ---
