+++ src/lib/api-grist.ts
@@ -1,75 +1,76 @@
 import { env } from 'next-runtime-env'
 
-const BASE_URL = env('NEXT_PUBLIC_GRIST_API_URL') || "";
-const DOC_ID = env('NEXT_PUBLIC_GRIST_DOC_ID') || "";
-const WANTED_TABLE_ID = env('NEXT_PUBLIC_GRIST_TABLE_ID') || "";
-const API_TOKEN = env('NEXT_PUBLIC_GRIST_API_TOKEN') || "";
+const BASE_URL = env('NEXT_PUBLIC_GRIST_API_URL') || ''
+const DOC_ID = env('NEXT_PUBLIC_GRIST_DOC_ID') || ''
+const WANTED_TABLE_ID = env('NEXT_PUBLIC_GRIST_TABLE_ID') || ''
+const API_TOKEN = env('NEXT_PUBLIC_GRIST_API_TOKEN') || ''
 
 interface GristRecord {
-  id: number;
+  id: number
   fields: {
-    [key: string]: any;
-  };
+    [key: string]: string
+  }
 }
 
 interface ProcessedRecord {
-  id_application: any;
-  nom_application: any;
-  description_utilisation: any;
-  nom_utilisateur: any;
-  image_utilisateur: any;
-  statut_integration: any;
-  type_integration: any;
-  url_application: any;
-  categorie_application: any;
-  tags_application: any;
+  id_application: string
+  nom_application: string
+  description_utilisation: string
+  nom_utilisateur: string
+  image_utilisateur: string
+  statut_integration: string
+  type_integration: string
+  url_article: string
+  url_application: string
+  categorie_application: string
+  tags_application: string
 }
 
 async function fetchTableJson(): Promise<{ records: GristRecord[] }> {
-  const filterDict = { "non_publication_usage": [false] };
-  const params = new URLSearchParams({ filter: JSON.stringify(filterDict) });
+  const filterDict = { non_publication_usage: [false] }
+  const params = new URLSearchParams({ filter: JSON.stringify(filterDict) })
 
   const response = await fetch(`${BASE_URL}/docs/${DOC_ID}/tables/${WANTED_TABLE_ID}/records?${params}`, {
     headers: {
-      "Authorization": `Bearer ${API_TOKEN}`,
+      Authorization: `Bearer ${API_TOKEN}`,
     },
     next: {
-      revalidate:0,
-      tags: ["grist"]
-    }
-  });
+      revalidate: 0,
+      tags: ['grist'],
+    },
+  })
 
   if (!response.ok) {
-    throw new Error(`Erreur HTTP ${response.status} lors de la récupération des données.`);
+    throw new Error(`Erreur HTTP ${response.status} lors de la récupération des données.`)
   }
 
-  return response.json();
+  return response.json()
 }
 
 function flattenTags(val: any): string {
   if (Array.isArray(val)) {
-    let cleaned = val;
-    if (cleaned.length > 1 && cleaned[0] === "L") {
-      cleaned = cleaned.slice(1);
+    let cleaned = val
+    if (cleaned.length > 1 && cleaned[0] === 'L') {
+      cleaned = cleaned.slice(1)
     }
-    cleaned = cleaned.filter((v: any) => v).map((v: any) => String(v));
-    return cleaned.join(", ");
+    cleaned = cleaned.filter((v: any) => v).map((v: any) => String(v))
+    return cleaned.join(', ')
   }
-  return val;
+  return val
 }
 
 export async function fetchAndProcessGristData(): Promise<ProcessedRecord[]> {
-  const data = await fetchTableJson();
-  const records = data.records || [];
+  const data = await fetchTableJson()
+  const records = data.records || []
 
   // Traiter les données
-  const processedRecords: ProcessedRecord[] = records.map(record => {
-    const fields = record.fields;
-    
+  const processedRecords: ProcessedRecord[] = records.map((record) => {
+    const fields = record.fields
+
     // Aplatir les tags
-    let tagsApplication = fields.tags_application;
+    let tagsApplication = fields.tags_application
     if (tagsApplication) {
-      tagsApplication = flattenTags(tagsApplication);
+      tagsApplication = flattenTags(tagsApplication)
     }
 
     return {
@@ -80,11 +81,12 @@ export async function fetchAndProcessGristData(): Promise<ProcessedRecord[]> {
       image_utilisateur: fields.image_utilisateur,
       statut_integration: fields.statut_integration,
       type_integration: fields.type_integration,
+      url_article: fields.url_article_blog,
       url_application: fields.url_application,
       categorie_application: fields.Thematique,
       tags_application: tagsApplication,
-    };
-  });
+    }
+  })
 
-  return processedRecords;
+  return processedRecords
 }
